requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
et-xmlfile==1.1.0
flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-swagger-ui==4.11.1
google-auth==2.27.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==8.5.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.27.0
numpy==1.24.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
tzdata==2023.4
urllib3==2.2.0
werkzeug==3.0.1
zipp==3.20.2